use the check_rate_limit from utils
